FOUR-14121 Create Page Rules and create svg file icons
