+++ resources/js/inbox-rules/components/PmRowButtons.vue
@@ -6,20 +6,16 @@
 
         <b-button variant="light"
                   size="sm"
-                  @click="$emit('onButtonEdit', row);">
-          <svg width="17" height="17" viewBox="0 0 17 17" fill="none" xmlns="http://www.w3.org/2000/svg">
-            <path d="M9.79965 3.42058L13.5763 7.19727L4.76229 16.0113H0.985596V12.2338L9.79965 3.42058ZM11.0586 2.1608L12.9469 0.27245C13.1139 0.105543 13.3403 0.0117798 13.5763 0.0117798C13.8124 0.0117798 14.0388 0.105543 14.2058 0.27245L16.7245 2.79024C16.8073 2.87293 16.8729 2.97112 16.9177 3.0792C16.9625 3.18728 16.9856 3.30314 16.9856 3.42014C16.9856 3.53714 16.9625 3.65299 16.9177 3.76107C16.8729 3.86916 16.8073 3.96735 16.7245 4.05003L14.8352 5.93838L11.0586 2.1608Z" fill="#6A7888"/>
-          </svg>
+                  @click="$emit('onEditRule', row);">
+          <img src="/img/pencil-fill.svg">
         </b-button>
 
         <div class="pm-row-buttons-separator"></div>
 
         <b-button variant="light"
                   size="sm"
-                  @click="$emit('onButtonRemove', row)">
-          <svg width="16" height="16" viewBox="0 0 16 16" fill="none" xmlns="http://www.w3.org/2000/svg">
-            <path d="M2 14.5001C2 14.8979 2.15803 15.2794 2.43934 15.5608C2.72064 15.8421 3.10217 16.0001 3.5 16.0001H12.5C12.8978 16.0001 13.2794 15.8421 13.5607 15.5608C13.842 15.2794 14 14.8979 14 14.5001V4.0001H2V14.5001ZM10.5 6.50009C10.5 6.36749 10.5527 6.24031 10.6464 6.14654C10.7402 6.05277 10.8674 6.00009 11 6.00009C11.1326 6.00009 11.2598 6.05277 11.3535 6.14654C11.4473 6.24031 11.5 6.36749 11.5 6.50009V13.5001C11.5 13.6327 11.4473 13.7599 11.3535 13.8536C11.2598 13.9474 11.1326 14.0001 11 14.0001C10.8674 14.0001 10.7402 13.9474 10.6464 13.8536C10.5527 13.7599 10.5 13.6327 10.5 13.5001V6.50009ZM7.5 6.50009C7.5 6.36749 7.55268 6.24031 7.64644 6.14654C7.74021 6.05277 7.86739 6.00009 8 6.00009C8.13261 6.00009 8.25978 6.05277 8.35355 6.14654C8.44732 6.24031 8.5 6.36749 8.5 6.50009V13.5001C8.5 13.6327 8.44732 13.7599 8.35355 13.8536C8.25978 13.9474 8.13261 14.0001 8 14.0001C7.86739 14.0001 7.74021 13.9474 7.64644 13.8536C7.55268 13.7599 7.5 13.6327 7.5 13.5001V6.50009ZM4.5 6.50009C4.5 6.36749 4.55268 6.24031 4.64645 6.14654C4.74021 6.05277 4.86739 6.00009 5 6.00009C5.13261 6.00009 5.25978 6.05277 5.35355 6.14654C5.44732 6.24031 5.5 6.36749 5.5 6.50009V13.5001C5.5 13.6327 5.44732 13.7599 5.35355 13.8536C5.25978 13.9474 5.13261 14.0001 5 14.0001C4.86739 14.0001 4.74021 13.9474 4.64645 13.8536C4.55268 13.7599 4.5 13.6327 4.5 13.5001V6.50009ZM14.5 1.0001H10.75L10.4562 0.415722C10.394 0.290789 10.2982 0.185698 10.1795 0.112272C10.0608 0.0388452 9.92394 -3.07113e-06 9.78437 9.70317e-05H6.2125C6.07324 -0.000438297 5.93665 0.0382651 5.81837 0.111773C5.7001 0.185281 5.60492 0.29062 5.54375 0.415722L5.25 1.0001H1.5C1.36739 1.0001 1.24021 1.05278 1.14645 1.14654C1.05268 1.24031 1 1.36749 1 1.5001V2.5001C1 2.6327 1.05268 2.75988 1.14645 2.85365C1.24021 2.94742 1.36739 3.0001 1.5 3.0001H14.5C14.6326 3.0001 14.7598 2.94742 14.8535 2.85365C14.9473 2.75988 15 2.6327 15 2.5001V1.5001C15 1.36749 14.9473 1.24031 14.8535 1.14654C14.7598 1.05278 14.6326 1.0001 14.5 1.0001Z" fill="#6A7888"/>
-          </svg>
+                  @click="$emit('onRemoveRule', row)">
+          <img src="/img/trash-fill.svg">
         </b-button>
 
       </template>
@@ -36,12 +32,6 @@
     props: {
       value: null,
       row: null
-    },
-    data() {
-      return {
-      };
-    },
-    mounted() {
     },
     methods: {
       show() {
